Hedgehog.Extras.Internal.Process

Documentation

binDist

Arguments

∷ (HasCallStack, MonadIO m)

⇒ String
Package name

→ String
Environment variable pointing to the binary to run (used for error messages only)

→ m FilePath
Path to executable

Consult the "plan.json" generated by cabal to get the path to the executable corresponding. to a haskell package. It is assumed that the project has already been configured and the executable has been built. Throws an exception on failure.

liftIOAnnotated ∷ (HasCallStack, MonadIO m) ⇒ IO a → m a

planJsonFile ∷ String

Discover the location of the plan.json file.