Hedgehog.Extras.Test.MonadAssertion

Documentation

class Monad m ⇒ MonadAssertion m where

A monad that supports assertions, allowing for throwing and catching assertion failures.

Methods

throwAssertion ∷ Failure → m a

catchAssertion ∷ m a → (Failure → m a) → m a

tryAssertion ∷ MonadAssertion m ⇒ m a → m (Either Failure a)

Attempt to run a function that may assert, returning either a failure or the result of the assertion.

tryExceptAssertion ∷ MonadAssertion m ⇒ MonadCatch m ⇒ m a → m (TestResult a)

Attempt to run a function that may throw an exception, returning a TestResult that encapsulates the result, failure, or error.

getTestResult ∷ MonadThrow m ⇒ MonadError Failure m ⇒ TestResult a → m a

Extract the result from a TestResult, throwing an error if the result is a failure or an exception.