Hedgehog.Extras.Test.Prim

Documentation

failWithCustom ∷ MonadTest m ⇒ CallStack → Maybe Diff → String → m a

Takes a CallStack so the error can be rendered at the appropriate call site.

failMessage ∷ MonadTest m ⇒ CallStack → String → m a

Takes a CallStack so the error can be rendered at the appropriate call site.