Hedgehog.Extras.Test.ProcessIO

Documentation

execFlex ∷ HasCallStack ⇒ MonadIO m ⇒ String → String → [String] → m String

Create a process returning its stdout.

Being a flex function means that the environment determines how the process is launched.

When running in a nix environment, the envBin argument describes the environment variable that defines the binary to use to launch the process.

When running outside a nix environment, the pkgBin describes the name of the binary to launch via cabal exec.

procFlex

Arguments

∷ HasCallStack

⇒ MonadIO m

⇒ String
Cabal package name corresponding to the executable

→ String
Environment variable pointing to the binary to run

→ [String]
Arguments to the CLI command

→ m CreateProcess
Captured stdout

Create a CreateProcess describing how to start a process given the Cabal package name corresponding to the executable, an environment variable pointing to the executable, and an argument list.

The actual executable used will the one specified by the environment variable, but if the environment variable is not defined, it will be found instead by consulting the "plan.json" generated by cabal. It is assumed that the project has already been configured and the executable has been built.