Hedgehog.Extras.Test.Tripwire

Documentation

data Tripwire

Represents a tripwire which can be triggered only once

makeTripwire ∷ MonadIO m ⇒ m Tripwire

Creates a new tripwire

triggerTripwire ∷ HasCallStack ⇒ MonadIO m ⇒ Tripwire → m ()

Triggers the tripwire and registers the place of the first trigger. Idempotent. Does not do do anything besides just registering the place where this function is called.

resetTripwire ∷ MonadIO m ⇒ Tripwire → m ()

Restore tripwire to initial non triggered state

checkTripwire ∷ HasCallStack ⇒ MonadTest m ⇒ MonadIO m ⇒ Tripwire → m ()

Fails the test if the tripwire was triggered. Prints the callstack where the tripwire was triggered.

isTriggered ∷ MonadIO m ⇒ Tripwire → m (Maybe CallStack)

Check if the tripwire is triggered. Return the first trigger location.