Hedgehog.Extras.Unit.Workspace

Documentation

unitWorkspace ∷ MonadTest m ⇒ MonadIO m ⇒ HasCallStack ⇒ MonadBaseControl IO m ⇒ FilePath → (FilePath → m ()) → m ()

Create a workspace directory which will exist for at least the duration of the supplied block.

The directory will have the supplied prefix but contain a generated random suffix to prevent interference between tests

The directory will be deleted if the block succeeds, but left behind if the block fails.

unitModuleWorkspace ∷ HasCallStack ⇒ MonadBaseControl IO m ⇒ MonadIO m ⇒ MonadTest m ⇒ String → (FilePath → m ()) → m ()

Create a workspace directory which will exist for at least the duration of the supplied block.

The directory will have the prefix as "$prefixPath/$moduleName" but contain a generated random suffix to prevent interference between tests

The directory will be deleted if the block succeeds, but left behind if the block fails.

The prefix argument should not contain directory delimeters.