Hedgehog.Extras.Test.File

Documentation

createDirectoryIfMissing ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m FilePath

Create the directory directory if it is missing.

createDirectoryIfMissing_ ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m ()

Create the directory directory if it is missing.

createSubdirectoryIfMissing ∷ HasCallStack ⇒ MonadTest m ⇒ MonadIO m ⇒ FilePath → FilePath → m FilePath

Create the subdirectory subdirectory if it is missing. The subdirectory is returned.

createSubdirectoryIfMissing_ ∷ HasCallStack ⇒ MonadTest m ⇒ MonadIO m ⇒ FilePath → FilePath → m ()

Create the subdirectory subdirectory if it is missing. The subdirectory is returned.

copyFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → FilePath → m ()

Copy the contents of the src file to the dst file.

renameFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → FilePath → m ()

Rename the src file to dst.

createFileLink ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → FilePath → m ()

Create a symbolic link from dst to src.

listDirectory ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m [FilePath]

List p directory.

appendFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → String → m ()

Append contents to the filePath file.

writeFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → String → m ()

Write contents to the filePath file.

openFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → IOMode → m Handle

Open a handle to the filePath file.

readFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m String

Read the contents of the filePath file.

lbsWriteFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → ByteString → m ()

Write contents to the filePath file.

lbsReadFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m ByteString

Read the contents of the filePath file.

textWriteFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → Text → m ()

Write contents to the filePath file.

textReadFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m Text

Read the contents of the filePath file.

copyRewriteJsonFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → FilePath → (Value → Value) → m ()

Rewrite the filePath JSON file using the function f.

readJsonFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m (Either String Value)

Read the filePath file as JSON.

readJsonFileOk ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m Value

Read the filePath file as JSON. Same as readJsonFile but fails on error.

rewriteJsonFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → (Value → Value) → m ()

Rewrite the filePath JSON file using the function f.

rewriteLbsJson ∷ (MonadTest m, HasCallStack) ⇒ (Value → Value) → ByteString → m ByteString

copyRewriteYamlFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → FilePath → (Value → Value) → m ()

Rewrite the filePath YAML file using the function f.

readYamlFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m (Either ParseException Value)

Read the filePath file as YAML.

readYamlFileOk ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m Value

Read the filePath file as YAML. Same as readYamlFile but fails on error.

rewriteYamlFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → (Value → Value) → m ()

Rewrite the filePath YAML file using the function f.

rewriteLbsYaml ∷ (MonadTest m, HasCallStack) ⇒ (Value → Value) → ByteString → m ByteString

cat ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m ()

Annotate the contents of the filePath file.

assertIsJsonFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m ()

Assert the filePath can be parsed as JSON.

assertIsYamlFile ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m ()

Assert the filePath can be parsed as YAML.

assertFileExists ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m ()

Asserts that the given file exists.

assertFilesExist ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ [FilePath] → m ()

Asserts that all of the given files exist.

assertFileMissing ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m ()

Asserts that the given file is missing.

assertFilesMissing ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ [FilePath] → m ()

Asserts that all of the given files are missing.

assertFileOccurences ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ Int → String → FilePath → m ()

Assert the file contains the given number of occurrences of the given string

assertFileLines ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ (Int → Bool) → FilePath → m ()

Assert the file contains the given number of occurrences of the given string

assertEndsWithSingleNewline ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → m ()

Assert the file contains the given number of occurrences of the given string

appendFileTimeDelta ∷ (MonadTest m, MonadIO m, HasCallStack) ⇒ FilePath → UTCTime → m ()

Write contents to the filePath file.